TT fixed. Incremental hash not working yet, will be fixed later. Using xor 56 for opponent PSQT rather than .reverse

+++ move_generator.py
@@ -92,7 +92,7 @@ def get_scored_moves(engine, moves, tt_move):
     scored_moves = []
 
     for move in moves:
-        scored_moves.append((move, score_move(engine, move, tt_move)))
+        scored_moves.append((move, nb.int32(score_move(engine, move, tt_move))))
 
     return scored_moves
